chore(bin/generate-docs.js): refactor docs generation
